+++ src/ergotest/test_result.ts
@@ -60,14 +60,14 @@ export abstract class TestResult {
 	/**
 	 * Create a TestResult for a suite of tests.
 	 * @param {string|string[]} name The name of the test. Can be a list of names.
-	 * @param {TestResult[]} children The nested results of this suite.
+	 * @param {TestResult[]} tests The nested tests in this suite (can be test suites or individual test cases).
 	 * @param {string} [options.filename] The file that contained this suite (optional).
 	 * @param {TestMarkValue} [options.mark] Whether this suite was marked with `.skip`, `.only`, or nothing.
 	 * @returns {TestSuiteResult} The result.
 	 */
 	static suite(
 		name: string | string[],
-		children: TestResult[],
+		tests: TestResult[],
 		{
 			filename,
 			mark = TestMark.none
@@ -86,7 +86,7 @@ export abstract class TestResult {
 		]);
 
 		if (!Array.isArray(name)) name = [ name ];
-		return new TestSuiteResult(name, children, mark, filename);
+		return new TestSuiteResult(name, tests, mark, filename);
 	}
 
 	/**
@@ -219,8 +219,8 @@ export abstract class TestResult {
 
 	/**
 	 * @returns {TestCaseResult[]} All test results, with a mark (.only, etc.) that matches the requested marks,
-	 *   flattened into a single list. This includes suites; although the test results are all in a single list, and are
-	 *   filtered, any suites in the list still have all their children.
+	 *   flattened into a single list, including test suites. However, if you access the properties of the test suites,
+	 *   such as {@link TestSuiteResult.tests}, those properties won’t be filtered.
 	 */
 	abstract allMatchingMarks(...marks: TestMarkValue[]): TestResult[];
 
@@ -265,15 +265,15 @@ export class TestSuiteResult extends TestResult {
 	}
 
 	private readonly _name: string[];
-	private readonly _children: TestResult[];
+	private readonly _tests: TestResult[];
 	private readonly _mark: TestMarkValue;
 	private readonly _filename?: string;
 
 	/** Internal use only. (Use {@link TestResult.suite} instead.) */
-	constructor(name: string[], children: TestResult[], mark: TestMarkValue, filename?: string) {
+	constructor(name: string[], tests: TestResult[], mark: TestMarkValue, filename?: string) {
 		super();
 		this._name = name;
-		this._children = children;
+		this._tests = tests;
 		this._mark = mark;
 		this._filename = filename;
 	}
@@ -297,11 +297,10 @@ export class TestSuiteResult extends TestResult {
 	}
 
 	/**
-	 * @returns { TestResult[] } This suite's direct children, which can either be test case results or test suite
-	 *   results.
+	 * @returns { TestResult[] } The tests in this suite, which can either be test case results or test suite results.
 	 */
-	get children(): TestResult[] {
-		return this._children;
+	get tests(): TestResult[] {
+		return this._tests;
 	}
 
 	/**
@@ -349,7 +348,7 @@ export class TestSuiteResult extends TestResult {
 		ensure.signature(arguments, []);
 
 		const tests: TestCaseResult[] = [];
-		this._children.forEach((result: TestResult) => {
+		this._tests.forEach((result: TestResult) => {
 			result.allTests().forEach(subTest => tests.push(subTest));
 		});
 		return tests;
@@ -365,9 +364,9 @@ export class TestSuiteResult extends TestResult {
 	}
 
 	/**
-	 * @returns {TestCaseResult[]} All the marked test results (.only, etc.), not including results without marks, but
-	 *   including suites, flattened into a single list. Although the test results are all in a single list, any suites
-	 *   in the list still have all their children.
+	 * @returns {TestCaseResult[]} All test results, with a mark (.only, etc.) that matches the requested marks,
+	 *   flattened into a single list, including test suites. However, if you access the properties of the test suites,
+	 *   such as {@link TestSuiteResult.tests}, those properties won’t be filtered.
 	 */
 	allMarkedResults(): TestResult[] {
 		ensure.signature(arguments, []);
@@ -382,7 +381,7 @@ export class TestSuiteResult extends TestResult {
 
 		const results = new Set<TestResult>();
 		if (marks.includes(this.mark)) results.add(this);
-		this._children.forEach((result: TestResult) => {
+		this._tests.forEach((result: TestResult) => {
 			if (marks.includes(result.mark)) results.add(result);
 			result.allMatchingMarks.apply(result, marks).forEach(subResult => results.add(subResult));
 		});
@@ -444,7 +443,7 @@ export class TestSuiteResult extends TestResult {
 			name: this._name,
 			mark: this._mark,
 			filename: this._filename,
-			suite: this._children.map(test => test.serialize()),
+			suite: this._tests.map(test => test.serialize()),
 		};
 	}
 
@@ -452,10 +451,10 @@ export class TestSuiteResult extends TestResult {
 		if (!(that instanceof TestSuiteResult)) return false;
 		if (this._mark !== that._mark) return false;
 
-		if (this._children.length !== that._children.length) return false;
-		for (let i = 0; i < this._children.length; i++) {
-			const thisResult = this._children[i]!;
-			const thatResult = that._children[i]!;
+		if (this._tests.length !== that._tests.length) return false;
+		for (let i = 0; i < this._tests.length; i++) {
+			const thisResult = this._tests[i]!;
+			const thatResult = that._tests[i]!;
 			if (!thisResult.equals(thatResult)) return false;
 		}