+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Links to other documentation:
 
 To Document:
 * TestResult.suite signature change
+* TestResult.children --> TestResult.tests rename
 
 
 ## v0.11.x: Add optional 'actual' and 'expected' to assert.fail()